Requirements fixed
